Fix output type for LANGUAGE_SETTING in main function

+++ tests/setting_qt/src/main.cc
@@ -9,6 +9,6 @@ int main() {
         return 1;
     }
 
-    qDebug() << settings->value(LANGUAGE_SETTING);
+    qDebug() << settings->value(LANGUAGE_SETTING).toInt();
     return 0;
 }
